move functions inside
