+++ conan/cache/cache.py
@@ -47,21 +47,11 @@ class Cache:
         from conan.cache.recipe_layout import RecipeLayout
         return RecipeLayout(ref, cache=self, manager=self._locks_manager)
 
-    @staticmethod
-    def get_default_reference_path(ref: ConanFileReference) -> Optional[str]:
-        if ref.revision:
-            return ref.full_str().replace('@', '/').replace('#', '/').replace(':', '/')  # TODO: TBD
-        else:
-            return None
+    def remove_reference(self, ref: ConanFileReference):
+        pass
 
-    @staticmethod
-    def get_default_package_path(pref: PackageReference, folder: ConanFolders) -> Optional[str]:
-        if pref.revision:
-            package_folder = pref.full_str().replace('@', '/') \
-                .replace('#', '/').replace(':', '/')  # TODO: TBD
-            return os.path.join(package_folder, folder.name)
-        else:
-            return None
+    def remove_package(self, pref: PackageReference):
+        pass
 
     def _move_rrev(self, old_ref: ConanFileReference, new_ref: ConanFileReference,
                    move_reference_contents: bool = False) -> Optional[str]:
@@ -73,7 +63,7 @@ class Cache:
 
         if move_reference_contents:
             old_path = self._backend.try_get_reference_directory(new_ref)
-            new_path = self.get_default_reference_path(new_ref)
+            new_path = self._backend.get_default_reference_path(new_ref)
             self._backend.update_path(new_ref, new_path)
             if os.path.exists(old_path):
                 shutil.move(old_path, new_path)
@@ -86,7 +76,7 @@ class Cache:
         self._backend.update_prev(old_pref, new_pref)
         if move_package_contents:
             old_path = self._backend.try_get_package_directory(new_pref, folder)
-            new_path = self.get_default_package_path(new_pref, folder)
+            new_path = self._backend.get_default_package_path(new_pref, folder)
             self._backend.update_path(new_pref, new_path)
             if os.path.exists(old_path):
                 shutil.move(old_path, new_path)
